pyproject.toml
[tool.poetry]
name = "igvf-catalog"
version = "0.1.0"
description = ""
authors = ["Daniel Zeiberg <zeiberg.d@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
scipy = "^1.12.0"
tqdm = "^4.66.1"
joblib = "^1.3.2"
fire = "^0.5.0"
pyarrow = "^15.0.0"
ensembl-rest = "^0.3.4"
biopython = "^1.83"
seaborn = "^0.13.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
